[Chore] #28 - 미션 cell 수정
